+++ parac/__init__.py
@@ -34,8 +34,6 @@ __copyright__ = "Luna Klatzer"
 # module imports
 import logging as lib_logging
 
-import colorama as lib_colorama
-
 from . import abc
 from . import compiler
 from . import const
@@ -43,6 +41,7 @@ from . import exceptions
 from . import logging
 from . import preprocessor
 from . import util
+
 # local imports
 from .const import *
 from .exceptions import *
@@ -73,7 +72,7 @@ __all__ = [
     *MODULES
 ]
 
-lib_colorama.init(autoreset=True)
+
 lib_logging.getLogger(__name__).addHandler(lib_logging.NullHandler())
 
 # An instance of the compiler, which should be generally used in the module